giddyup/pyproject.toml
[project]
name = "giddyup"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Sean Monaghan", email = "SeanMonaghan@bitbucket.org" }
]
requires-python = ">=3.13"
dependencies = [
    "connectorx>=0.4.4",
    "duckdb>=1.4.1",
    "hydra-core>=1.3.2",
    "lightgbm>=4.6.0",
    "mlflow>=3.5.0",
    "polars>=1.34.0",
    "prefect>=3.4.24",
    "psycopg[binary]>=3.2.10",
    "pyarrow>=21.0.0",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "sqlalchemy>=2.0.44",
    "xgboost>=3.0.5",
]

[project.scripts]
giddyup = "giddyup:main"

[build-system]
requires = ["uv_build>=0.9.3,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "ruff>=0.14.1",
]
